+++ qcell/gui/mixin_palette.py
@@ -100,6 +100,7 @@ class PaletteMixin:
             "Terminal...": self.show_terminal,
             "Matrix tool...": self.show_matrix_tool,
             "RF toolkit...": self.show_rf_tool,
+            "Smith chart...": self.show_smith_chart,
             "Python console...": self.show_pyconsole,
             "Graph...": self.show_graph,
             "Equation editor...": self.show_equation,